Starts with the implementation of swiping
After jquery-mobile was added to the scripts, a new problem occured.
jQuery-mobile automatically removes hashtages from the URL, so it wasn't
possible to redirect to viewing the article. I partially fixed this
issue, by which I mean that it is possible to view an article, but the
link has to be clicked twice.